% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{apply_agg_fun}
\alias{apply_agg_fun}
\title{Function aggregate mw data with a given temporal resolution and function}
\usage{
apply_agg_fun(mw, agg_fun, agg_time)
}
\arguments{
\item{mw}{dbplyr database connection tbl}

\item{agg_fun}{character, name of the aggregation function - may either be "mean", "min", "max" or "none"}

\item{agg_time}{numeric, temporal resolution of the aggregation}
}
\description{
Function aggregate mw data with a given temporal resolution and function
}
