% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{canConnect}
\alias{canConnect}
\title{Checks whether login information is valid
Wrapper for frwfutils::dbCanConnect}
\usage{
canConnect(login_credentials)
}
\arguments{
\item{login_credentials}{data.frame, contains login information}
}
\description{
Checks whether login information is valid
Wrapper for frwfutils::dbCanConnect
}
