% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{connect}
\alias{connect}
\title{Connects to database via data.frame
Wrapper for frwfutils::db_connect}
\usage{
connect(login_credentials)
}
\arguments{
\item{login_credentials}{data.frame, contains login information}
}
\description{
Connects to database via data.frame
Wrapper for frwfutils::db_connect
}
