% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{download_LI}
\alias{download_LI}
\title{download LI datatable with filter
connects to database to download LI table with filter and disconnects after finishing or error}
\usage{
download_LI(Stat, login_credentials)
}
\arguments{
\item{Stat}{integer, id of the stat}

\item{login_credentials}{data.frame, contains login information}
}
\description{
download LI datatable with filter
connects to database to download LI table with filter and disconnects after finishing or error
}
