% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{download_MW}
\alias{download_MW}
\title{download information on which Spots, Paramaters, Messpositions exist for a given Stat and time frame
connects to database  to get information and disconnects after finishing or error}
\usage{
download_MW(
  Stat,
  Spot = -999,
  Para = -999,
  Messposition = -999,
  start_date = Sys.time() - 7 * 86400,
  end_date = Sys.time(),
  Quali = -999,
  login_credentials
)
}
\arguments{
\item{Stat}{integer, id of the stat}

\item{Spot}{integer, if of the measurement location}

\item{Para}{integer, id of the parameter}

\item{Messposition}{integer, id of vertical location of measurement, can be above or below ground}

\item{start_date}{posixct, lower range of the time frame}

\item{end_date}{posixct, upper range of the time frame}

\item{Quali, }{integer, id value of dataqualitaet (1 = raw data)}

\item{login_credentials}{data.frame, contains login information}
}
\description{
download information on which Spots, Paramaters, Messpositions exist for a given Stat and time frame
connects to database  to get information and disconnects after finishing or error
}
