% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{download_static_tables}
\alias{download_static_tables}
\title{download tables that contain static information
connects to databaseto get informationr and disconnects after finishing or error}
\usage{
download_static_tables(login_credentials)
}
\arguments{
\item{login_credentials}{data.frame, contains login information}
}
\description{
download tables that contain static information
connects to databaseto get informationr and disconnects after finishing or error
}
