% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{full_download}
\alias{full_download}
\title{download datatable
connects to database to download all content of a table and disconnects after finishing or error}
\usage{
full_download(table_name = "Tab_Stat", login_credentials)
}
\arguments{
\item{table_name}{character, name of the database table}

\item{login_credentials}{data.frame, contains login information}
}
\description{
download datatable
connects to database to download all content of a table and disconnects after finishing or error
}
