% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_health}
\alias{plot_health}
\title{plot time passed since last measurement for all devices}
\usage{
plot_health(last_signal, stbls)
}
\arguments{
\item{last_signal}{data.frame, download result by download_last_signal}

\item{stbls}{list, download result by download_static_tables}
}
\value{
ggplot2 object
}
\description{
plot time passed since last measurement for all devices
}
