% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_last_value}
\alias{plot_last_value}
\title{plot value of the last measurement for all active devices}
\usage{
plot_last_value(last_signal, stbls)
}
\arguments{
\item{last_signal}{data.frame, download result by download_last_signal}

\item{stbls}{list, download result by download_static_tables}
}
\value{
ggiraph object
}
\description{
plot value of the last measurement for all active devices
}
