% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_mw}
\alias{plot_mw}
\title{plot time series of measurements}
\usage{
plot_mw(mw, mw_info, input, agg_name)
}
\arguments{
\item{mw}{data.frame, download result by download_mw}

\item{input}{list, shiny input}

\item{agg_name, }{character, name of the current aggregation to adjust y-axis label}
}
\value{
ggplot2 object
}
\description{
plot time series of measurements
}
