% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_utils.R
\name{download_MW_info}
\alias{download_MW_info}
\title{download information on which Spots, Paramaters, Messpositions exist for a given Stat and time frame
connects to database to get information and disconnects after finishing or error}
\usage{
download_MW_info(
  Stat,
  Spot = -999,
  start_date = Sys.time() - 7 * 86400,
  end_date = Sys.time(),
  login_credentials
)
}
\arguments{
\item{Stat}{character, name of the stat}

\item{Spot}{character, name of the measurement locaiton}

\item{start_date}{posixct, lower range of the time frame}

\item{end_date}{posixct, upper range of the time frame}

\item{login_credentials}{data.frame, contains login information
#'}
}
\description{
download information on which Spots, Paramaters, Messpositions exist for a given Stat and time frame
connects to database to get information and disconnects after finishing or error
}
